% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{l_vec_compute_arma}
\alias{l_vec_compute_arma}
\title{l.vec.cmpt}
\usage{
l_vec_compute_arma(k, l_vec, cn_vec, m)
}
\arguments{
\item{k}{interger}

\item{l_vec}{vector}

\item{cn_vec}{vector}

\item{m}{integer}
}
\description{
l.vec.cmpt
}
