% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pmn_mdfft_arma}
\alias{pmn_mdfft_arma}
\title{mdfft method}
\usage{
pmn_mdfft_arma(nnt, pp, nn_vec, l_vec, cn_vec)
}
\arguments{
\item{nnt}{interger}

\item{pp}{matrix}

\item{nn_vec}{vector}

\item{l_vec}{vector}

\item{cn_vec}{vector}
}
\value{
result
}
\description{
mdfft method
}
